% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmm.R
\name{plot_glmm}
\alias{plot_glmm}
\title{Title}
\usage{
plot_glmm(
  model,
  data,
  predictor,
  outcome,
  grouping_var,
  family = "gaussian",
  y_scale = NULL,
  x_limits = NULL,
  y_limits = NULL,
  x_label = NULL,
  y_label = NULL,
  plot_title = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  x_num_size = 10,
  y_num_size = 10
)
}
\arguments{
\item{model}{Your multilevel model, should be `lmer`, `glmer` or `glmmTMB`}

\item{data}{Your data}

\item{predictor}{The predictor you would like to plot, in string form}

\item{outcome}{The outcome of the model}

\item{grouping_var}{The grouping/cluster variable, in string form}

\item{family}{Declare your family distribution, default is `"gaussian"`}

\item{y_scale}{Declare your y-axis scale, default is `"response"` which is identical to the model response, e.g., `"odds ratio"` for binomial link}

\item{x_limits}{Set the x-axis minimum and maximum value}

\item{y_limits}{Set the y-axis minimum and maximum value}

\item{x_label}{Set the x-axis name}

\item{y_label}{Set the y-axis name}

\item{plot_title}{Set the plot title}

\item{x_breaks}{Set the breaks of x-axis}

\item{y_breaks}{Set the breaks of y-axis}

\item{x_num_size}{Set the x-axis value size}

\item{y_num_size}{Set the y-axis value size}
}
\value{
a plot
}
\description{
Title
}
